src/requirements.txt
chromadb-client==0.6.3